gateway/bme280_data_fetcher/requirements.txt
paho-mqtt==1.4.0
RPi.bme280==0.2.3
smbus2==0.2.3
